refactor: update dir
