+++ packages/element3/src/utils/popupService/popup-manager.js
@@ -1,4 +1,4 @@
-import { addClass, removeClass } from '../../utils/dom'
+import { addClass, removeClass } from '../dom'
 
 let hasModal = false
 let hasInitZIndex = false
